+++ packages/relay/main.go
@@ -14,8 +14,8 @@ func main() {
 	signal.Notify(stopCh, os.Interrupt, syscall.SIGTERM)
 
 	// Get flags and log them
-	relay.InitFlags()
-	relay.GetFlags().DebugLog()
+	relay.InitRelayFlags()
+	relay.GetRelayFlags().DebugLog()
 
 	// Init WebRTC API
 	err = relay.InitWebRTCAPI()
